backend/requirements.txt
flask
flask-cors
psycopg2-binary
bcrypt
PyJWT
yfinance
numpy
pandas
ta
scikit-learn
joblib
tensorflow

### Install TA-Lib manually if you want its extra indicators (optional).
### This project uses the pure-Python `ta` package above, which is enough
### for the shipped predictor — TA-Lib is NOT required.

# Mac:        brew install ta-lib && pip install TA-Lib
# Ubuntu/WSL: sudo apt-get install ta-lib && pip install TA-Lib
# Windows:    download wheel from https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib
